api/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.0
gunicorn>=21.2
whitenoise>=6.6
dj-database-url>=2.2
python-dotenv>=1.0
psycopg[binary]>=3.2
requests>=2.31